OT132-18 Minor correction: change data tipe integer to long. Correct notation in nullable attributes

+++ src/main/java/com/alkemy/ong/repository/ICategoryRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ICategoryRepository extends JpaRepository<Category,Integer>{
+public interface ICategoryRepository extends JpaRepository<Category,Long>{
     
 }
